backend/requirements.txt
Flask
gunicorn
pymongo
python-dotenv
flask-cors